Add await support to BT
`BTRoot` has been modified with an additional exported variable named
`allow_await`. When `true`, the `BTRoot` modifies its flow such that it
`await`s for `tick()`, and suspends processing until that `await`
returns.

This can be useful when wanting to truly have an actor wait for
something to happen without having to fudge around it with nested FSMs
or other hacks.